Amend the readme and checkstyle config
