+++ src/test/java/edu/wpi/first/outlineviewer/model/NetworkTableBooleanTest.java
@@ -1,0 +1,16 @@
+package edu.wpi.first.outlineviewer.model;
+
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+
+public class NetworkTableBooleanTest {
+
+  @Test
+  public void testNetworkTableBoolean() {
+    final boolean value = false;
+    NetworkTableData data = NetworkTableData.createNetworkTableData("", value);
+
+    assertEquals(value, ((NetworkTableBoolean) data).valueProperty().getValue());
+  }
+}
